New TLS allocation function.
We use htscodecs_tls_alloc and htscodecs_tls_free to allocate and
deallocate a few specific blocks of memory that we would otherwise
hold on the stack.  Small stack size forbids us from doing this, but
equally so using malloc/free can sometimes carry a high penalty and
excess context switches to the kernel and back.

We previously solved this by Thread Local Storage.  This is the same
thing, but wrapped up in a more concise interface.

Note this is *not* a general purpose allocator and it should not be
used outside of this library.

As an indication of why this is necessary, here are some timings for
converting 10 million CRAM 3.1 records to 3.1 "small" profile using
scramble:

V3.1 norm to V3.1 small, with TLS:
real    0m23.632s       real    0m23.577s
user    1m44.174s       user    1m44.321s
sys     0m0.588s        sys     0m0.744s

V3.1 norm to V3.1 small, with -DNO_THREADS:
real    0m28.090s       real    0m27.766s
user    1m45.571s       user    1m45.758s
sys     0m16.231s       sys     0m15.174s

The system time difference is mainly down to mmap/munmap.  It can be
partially avoided using libtcmalloc, but then we have a cost of calloc
over malloc and higher user CPU time.  The reason we can avoid this
with TLS is we only calloc on the first call and (mostly) reuse
without zeroing after.  (This is as we can be sure the memory we're
reusing can never contain sensitive information.)