Combine filling SIMD lookup tables with RansEncSymbolInit()
Only entries that are used get initialised in syms, so only
the corresponding entries in the SA, SB, SC and SD arrays need
to be set.  It makes sense to do both at the same time.

Fixes an "illegal shift of negative value" complaint from UBSAN
due to reading the unused syms[] entries.