Fix UBSAN warnings about operations on NULL pointers
When decoding bit-packed data in the arithmetic codec, or RLE in
rANS, don't try to set tmp1 = NULL if there is no input data to
avoid later complaints from UBSAN about use of NULL pointers.
Setting tmp_size = 0 in these cases is sufficient to stop
the decoders from using any of the input.